Windows Certificate Store 관련 코드들을 별도 모듈로 분리
