+++ MonitoringSystem/MonitoringSystem.vcxproj.filters
@@ -24,5 +24,13 @@
     <ClCompile Include="..\Common\TcpSendRecvJpeg.cpp">
       <Filter>소스 파일</Filter>
     </ClCompile>
+    <ClCompile Include="WindowsKeyStoreAdapter.cpp">
+      <Filter>소스 파일</Filter>
+    </ClCompile>
+  </ItemGroup>
+  <ItemGroup>
+    <ClInclude Include="WindowsKeyStoreAdapter.h">
+      <Filter>헤더 파일</Filter>
+    </ClInclude>
   </ItemGroup>
 </Project>
