+++ MonitoringSystem/WindowsKeyStoreAdapter.h
@@ -1,0 +1,5 @@
+#pragma once
+
+#include <openssl/ssl.h>
+
+bool loadCertificatesFromWCS(SSL_CTX* ctx);
